Update contributors page
