--- frontend/src/Pages/ContributorsPage/components/index.ts
@@ -1,1 +1,0 @@
-export { type Contributor, ContributorItem } from './Contributor';
