+++ frontend/src/Pages/ContributorsPage/types.ts
@@ -1,0 +1,11 @@
+export type Contributor = {
+  name: string;
+  from: string;
+  to?: string;
+  github?: string;
+  picture?: string;
+  websjef?: {
+    from: string;
+    to: string;
+  };
+};
